+++ python/src/types/state.rs
@@ -1,0 +1,35 @@
+use pyo3::prelude::*;
+
+use opening_hours_syntax::rules::RuleKind;
+
+/// Specify the state of an opening hours interval.
+#[pyclass(ord, eq, frozen, hash, str, rename_all = "UPPERCASE")]
+#[derive(Hash, PartialOrd, Ord, PartialEq, Eq)]
+pub enum State {
+    /// Currently open
+    Open,
+    /// Currently closed
+    Closed,
+    /// May be open depending on context
+    Unknown,
+}
+
+impl From<RuleKind> for State {
+    fn from(kind: RuleKind) -> Self {
+        match kind {
+            RuleKind::Open => Self::Open,
+            RuleKind::Closed => Self::Closed,
+            RuleKind::Unknown => Self::Unknown,
+        }
+    }
+}
+
+impl std::fmt::Display for State {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            State::Open => write!(f, "open"),
+            State::Closed => write!(f, "closed"),
+            State::Unknown => write!(f, "unknown"),
+        }
+    }
+}